C++0x compatibility (fix narrowing conversions in initialisers; avoid ambiguity with std::next)
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@7739 3db68df2-c116-0410-a063-a993310a9797

+++ source/lib/res/graphics/unifont.cpp
@@ -141,7 +141,7 @@ static LibError UniFont_reload(UniFont* f, const PIVFS& vfs, const VfsPath& base
 		GLfloat w = (GLfloat)Width  / (GLfloat)TextureWidth;
 		GLfloat h = (GLfloat)Height / (GLfloat)TextureHeight;
 
-		GlyphData g = { u, -v, u+w, -v+h, OffsetX, -OffsetY, OffsetX+Width, -OffsetY+Height, Advance };
+		GlyphData g = { u, -v, u+w, -v+h, (i16)OffsetX, (i16)-OffsetY, (i16)(OffsetX+Width), (i16)(-OffsetY+Height), (i16)Advance };
 		(*f->glyphs)[(u16)Codepoint] = g;
 	}
